% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{in_viewer}
\alias{in_viewer}
\title{Display email message in RStudio viewer pane.}
\usage{
in_viewer(x)
}
\arguments{
\item{x}{message}
}
\value{
An \link[rstudioapi:viewer]{rstudioapi::viewer} object. Called for the side effect of opening the viewer pane and displays message
}
\description{
Only opens the viewer pane if interactive and RStudio is the IDE, otherwise returns the input invisibly.
}
\keyword{internal}
