% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mail_merge.R
\name{mm_mail_merge}
\alias{mm_mail_merge}
\title{Send course email.}
\usage{
mm_mail_merge(delegates, message, preview = TRUE, draft = TRUE,
  sleep = 1)
}
\arguments{
\item{delegates}{A \code{tibble} with all the columns that should be glued into
the message}

\item{message}{A list with components \code{yaml} and \code{body}.  See
\code{\link[=parse_email_from_googledoc]{parse_email_from_googledoc()}}}

\item{preview}{If \code{TRUE} displays message in viewer without sending mail}

\item{draft}{if \code{TRUE} sends a draft, otherwise sends the real email}

\item{sleep}{If \code{draft == TRUE} the number of seconds to sleep between each
preview}
}
\description{
Send course email.
}
