% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_parse_email.R
\name{mm_parse_email}
\alias{mm_parse_email}
\title{Parse a markdown document into yaml and body components.}
\usage{
mm_parse_email(txt)
}
\arguments{
\item{txt}{R Markdown document}
}
\value{
A list with components \code{yaml} and \code{body}
}
\description{
Parse a markdown document into yaml and body components.
}
