% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledocs.R
\name{mm_parse_email_from_googledoc}
\alias{mm_parse_email_from_googledoc}
\title{Parse the email inside a markdown document on google drive}
\usage{
mm_parse_email_from_googledoc(id)
}
\arguments{
\item{id}{Unique Google Drive identifier}
}
\description{
Parse the email inside a markdown document on google drive
}
