% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm_preview_mail}
\alias{mm_preview_mail}
\title{Preview message in viewer pane.}
\usage{
mm_preview_mail(to, subject = NULL, body, cc = NULL)
}
\arguments{
\item{to}{A data frame with one row, containing information about the student.}

\item{subject}{The email subject.}

\item{body}{The email message, in markdown format.}

\item{cc}{The email cc.}
}
\value{
HTML text
}
\description{
Preview message in viewer pane.
}
