% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledocs.R
\name{mm_read_googledoc}
\alias{mm_read_googledoc}
\title{Read a markdown file from google drive}
\usage{
mm_read_googledoc(id)
}
\arguments{
\item{id}{Unique Google Drive identifier}
}
\value{
the imported document
}
\description{
Read a markdown file from google drive
}
