% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_email.R
\name{mm_read_message}
\alias{mm_read_message}
\title{Parse a markdown document into yaml and body components.}
\usage{
mm_read_message(txt)
}
\arguments{
\item{txt}{A file in markdown format}
}
\value{
A list with components \code{yaml} and \code{body}
}
\description{
You can use this to construct the \code{message} argument of \code{\link[=mail_merge]{mail_merge()}}.
}
\seealso{
Other parsing functions: 
\code{\link{mm_read_message_googledoc}()}
}
\concept{parsing functions}
