% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm_send_draft}
\alias{mm_send_draft}
\title{Send draft email.}
\usage{
mm_send_draft(delegate, body, subject, draft = TRUE)
}
\arguments{
\item{delegate}{A data frame with one row, containing information about the student.}

\item{body}{The email message, in markdown format.}

\item{subject}{Subject line}

\item{draft}{if \code{TRUE} sends a draft, otherwise sends the real email}
}
\value{
A draft email message
}
\description{
Send draft email.
}
