% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm_send_mail}
\alias{mm_send_mail}
\title{Send email.}
\usage{
mm_send_mail(delegate, body, subject, draft = FALSE)
}
\arguments{
\item{delegate}{A data frame with one row, containing information about the student.}

\item{body}{The email message, in markdown format.}

\item{subject}{Subject line}

\item{draft}{if \code{TRUE} sends a draft, otherwise sends the real email}
}
\value{
An email message
}
\description{
Send email.
}
