% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{preview_mailmerge}
\alias{preview_mailmerge}
\title{Preview mailmerge as shiny gadget in RStudio preview pane.}
\usage{
preview_mailmerge(x)
}
\arguments{
\item{x}{mailmer_preview object}
}
\description{
Preview mailmerge as shiny gadget in RStudio preview pane.
}
