% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesno.R
\name{yesno}
\alias{yesno}
\title{Wait for interactive response to yes/no challenge question.}
\usage{
yesno(...)
}
\arguments{
\item{...}{Arguments to paste together to form challenge question.}
}
\value{
logical value
}
\description{
Copied from \code{devtools:::yesno}
}
