% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{in_viewer}
\alias{in_viewer}
\title{Display email message in RStudio viewer pane}
\usage{
in_viewer(x)
}
\arguments{
\item{x}{message}
}
\value{
opens the viewer pane and displays message
}
\description{
Display email message in RStudio viewer pane
}
