% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_merge.R
\name{mail_merge}
\alias{mail_merge}
\title{Send course email.}
\usage{
mail_merge(
  data,
  message,
  preview = TRUE,
  draft = TRUE,
  sleep_preview = 1,
  sleep_send = 0.1
)
}
\arguments{
\item{data}{A \code{tibble} with all the columns that should be glued into
the message}

\item{message}{A list with components \code{yaml} and \code{body}.  See
\code{\link[=mm_parse_email_from_googledoc]{mm_parse_email_from_googledoc()}}}

\item{preview}{If \code{TRUE} displays message in viewer without sending mail}

\item{draft}{if \code{TRUE} sends a draft, otherwise sends the real email}

\item{sleep_preview}{If \code{draft == TRUE} the number of seconds to sleep between each
preview}

\item{sleep_send}{If \code{draft == FALSE} the number of seconds to sleep between
each email send (to prevent gmail API 500 errors)}
}
\description{
Send course email.
}
