% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm_preview_mail}
\alias{mm_preview_mail}
\title{Preview message in viewer pane.}
\usage{
mm_preview_mail(delegate, body)
}
\arguments{
\item{delegate}{A data frame with one row, containing information about the student.}

\item{body}{The email message, in markdown format.}
}
\value{
HTML text
}
\description{
Preview message in viewer pane.
}
