% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm_preview_mail}
\alias{mm_preview_mail}
\title{Preview message in viewer pane.}
\usage{
mm_preview_mail(to, subject = "", body, cc = "")
}
\arguments{
\item{to}{The email \code{to} address}

\item{subject}{The email subject}

\item{body}{The email message, in markdown format}

\item{cc}{The email \code{cc} address}
}
\value{
HTML text
}
\description{
Preview message in viewer pane.
}
\keyword{internal}
