% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledocs.R
\name{mm_read_googledoc}
\alias{mm_read_googledoc}
\title{Read a markdown file from google drive}
\usage{
mm_read_googledoc(id)
}
\arguments{
\item{id}{Unique Google Drive identifier, passed to \link[googledrive:drive_download]{googledrive::drive_download}}
}
\value{
the imported document

A list of character strings, containing the content of the google doc
}
\description{
Read a markdown file from google drive
}
