% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm_send_draft}
\alias{mm_send_draft}
\title{Send draft email.}
\usage{
mm_send_draft(to, body, subject, cc = NULL, draft = TRUE)
}
\arguments{
\item{to}{The email \code{to} address}

\item{body}{The email message, in markdown format}

\item{subject}{The email subject}

\item{cc}{The email \code{cc} address}

\item{draft}{if \code{TRUE} sends a draft, otherwise sends the real email}
}
\value{
A draft email message
}
\description{
Send draft email.
}
\keyword{internal}
