% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm_send_mail}
\alias{mm_send_mail}
\title{Send email.}
\usage{
mm_send_mail(
  to,
  body,
  subject,
  cc = NULL,
  draft = FALSE,
  test = Sys.getenv("mailmerge_test", FALSE)
)
}
\arguments{
\item{to}{The email \code{to} address}

\item{body}{The email message, in markdown format}

\item{subject}{The email subject}

\item{cc}{The email \code{cc} address}

\item{draft}{if \code{TRUE} sends a draft, otherwise sends the real email}

\item{test}{If TRUE, doesn't send email but simply returns \code{mime} value, used for testing.}
}
\value{
An email message
}
\description{
Send email.
}
\keyword{internal}
